chunked: split file operations to a new file
